requirements.txt
pyyaml>=6.0
python-dotenv>=1.0
discord.py>=2.3
flask>=3.0
pytest>=8.0
